# crates/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HubBoard": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/HubColumn"
            },
            "type": "array"
          },
          "context": {
            "$ref": "#/components/schemas/HubBoardContext"
          },
          "orphans": {
            "$ref": "#/components/schemas/HubOrphans"
          }
        },
        "required": [
          "columns",
          "orphans",
          "context"
        ],
        "type": "object"
      },
      "HubBoardContext": {
        "properties": {
          "generatedAt": {
            "type": "string"
          },
          "jiraProjectKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "repos": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "repos",
          "generatedAt"
        ],
        "type": "object"
      },
      "HubBrief": {
        "properties": {
          "generatedAt": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "summary",
          "generatedAt"
        ],
        "type": "object"
      },
      "HubColumn": {
        "properties": {
          "id": {
            "type": "string"
          },
          "jiraStatuses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/HubTicketCard"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "jiraStatuses",
          "tickets"
        ],
        "type": "object"
      },
      "HubDeployment": {
        "properties": {
          "environment": {
            "type": "string"
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "environment"
        ],
        "type": "object"
      },
      "HubFavoriteWorkflow": {
        "properties": {
          "lastRun": {
            "description": "ISO-8601 millis of the favorite workflow's most recent run, or null.",
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          },
          "workflowId": {
            "type": "string"
          },
          "workflowName": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "workflowId",
          "workflowName",
          "path"
        ],
        "type": "object"
      },
      "HubInbox": {
        "properties": {
          "brief": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubBrief"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HubInboxItem"
            },
            "type": "array"
          },
          "rankedBy": {
            "description": "`ai` | `heuristic`",
            "type": "string"
          }
        },
        "required": [
          "items",
          "rankedBy"
        ],
        "type": "object"
      },
      "HubInboxItem": {
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "description": "`qa` | `review` | `run` | `reminder`",
            "type": "string"
          },
          "occurredAt": {
            "type": "string"
          },
          "qa": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubInboxQa"
              }
            ]
          },
          "reminder": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Reminder"
              }
            ]
          },
          "review": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubInboxReview"
              }
            ]
          },
          "run": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubInboxRun"
              }
            ]
          },
          "sourceLabel": {
            "type": "string"
          },
          "ticketKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "urgency": {
            "description": "Rank position as a string, ascending (`\"0\"` = most urgent).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "sourceLabel",
          "title",
          "occurredAt",
          "urgency"
        ],
        "type": "object"
      },
      "HubInboxQa": {
        "properties": {
          "channelId": {
            "type": "string"
          },
          "permalink": {
            "type": [
              "string",
              "null"
            ]
          },
          "threadTs": {
            "type": "string"
          },
          "unread": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "channelId",
          "threadTs",
          "unread"
        ],
        "type": "object"
      },
      "HubInboxReview": {
        "properties": {
          "prNumber": {
            "format": "int64",
            "type": "integer"
          },
          "repo": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "prNumber",
          "title",
          "url"
        ],
        "type": "object"
      },
      "HubInboxRun": {
        "properties": {
          "conclusion": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "workflowName": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "runId",
          "workflowName",
          "conclusion",
          "url"
        ],
        "type": "object"
      },
      "HubLinkBody": {
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "prNumber": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "repo": {
            "description": "`owner/name`",
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "ticketKey",
          "repo"
        ],
        "type": "object"
      },
      "HubLinkSuggestion": {
        "properties": {
          "confidence": {
            "format": "double",
            "type": "number"
          },
          "source": {
            "description": "matcher that produced the suggestion — currently always `heuristic`",
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "ticketKey",
          "confidence",
          "source"
        ],
        "type": "object"
      },
      "HubOrphanBranch": {
        "properties": {
          "name": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          },
          "suggestion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubLinkSuggestion"
              }
            ]
          }
        },
        "required": [
          "repo",
          "name"
        ],
        "type": "object"
      },
      "HubOrphanPr": {
        "properties": {
          "branch": {
            "type": "string"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "repo": {
            "type": "string"
          },
          "suggestion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubLinkSuggestion"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "number",
          "title",
          "branch",
          "url"
        ],
        "type": "object"
      },
      "HubOrphanTicket": {
        "properties": {
          "key": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "summary",
          "status"
        ],
        "type": "object"
      },
      "HubOrphans": {
        "properties": {
          "branches": {
            "items": {
              "$ref": "#/components/schemas/HubOrphanBranch"
            },
            "type": "array"
          },
          "prs": {
            "items": {
              "$ref": "#/components/schemas/HubOrphanPr"
            },
            "type": "array"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/HubOrphanTicket"
            },
            "type": "array"
          }
        },
        "required": [
          "prs",
          "branches",
          "tickets"
        ],
        "type": "object"
      },
      "HubPrRef": {
        "properties": {
          "branch": {
            "type": "string"
          },
          "checkState": {
            "description": "`success` | `running` | `failed` | `none`",
            "type": "string"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "repo": {
            "description": "`owner/name`",
            "type": "string"
          },
          "state": {
            "description": "`open` | `draft` | `merged` | `closed`",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "number",
          "title",
          "state",
          "checkState",
          "branch",
          "url"
        ],
        "type": "object"
      },
      "HubRunPill": {
        "properties": {
          "isFavorite": {
            "type": "boolean"
          },
          "repo": {
            "description": "`owner/name`",
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "startedAt": {
            "type": "string"
          },
          "status": {
            "description": "`success` | `running` | `failed`",
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflowId": {
            "type": "string"
          },
          "workflowName": {
            "type": "string"
          }
        },
        "required": [
          "repo",
          "workflowId",
          "workflowName",
          "runId",
          "status",
          "startedAt",
          "url",
          "isFavorite"
        ],
        "type": "object"
      },
      "HubRuns": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/HubRunPill"
            },
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "type": "object"
      },
      "HubSyncSource": {
        "properties": {
          "asOf": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopeCount": {
            "format": "int64",
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "source",
          "state",
          "scopeCount"
        ],
        "type": "object"
      },
      "HubSyncStatus": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/HubSyncSource"
            },
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "type": "object"
      },
      "HubTicketCard": {
        "properties": {
          "assignee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JiraUser"
              }
            ]
          },
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "bugCount": {
            "format": "int64",
            "type": "integer"
          },
          "checkState": {
            "description": "`success` | `running` | `failed` | `none` (aggregate over linked PRs)",
            "type": "string"
          },
          "deployment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubDeployment"
              }
            ]
          },
          "favoriteWorkflow": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HubFavoriteWorkflow"
              }
            ]
          },
          "key": {
            "type": "string"
          },
          "notesCount": {
            "format": "int64",
            "type": "integer"
          },
          "prs": {
            "items": {
              "$ref": "#/components/schemas/HubPrRef"
            },
            "type": "array"
          },
          "ready": {
            "type": "boolean"
          },
          "remindersDue": {
            "format": "int64",
            "type": "integer"
          },
          "slackUnread": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "summary",
          "status",
          "updatedAt",
          "bugCount",
          "ready",
          "prs",
          "checkState",
          "slackUnread",
          "notesCount",
          "remindersDue"
        ],
        "type": "object"
      },
...
      "JiraUser": {
        "properties": {
          "accountId": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "emailAddress": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "accountId",
          "displayName"
        ],
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "Reminder": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "dueAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "snoozedUntil": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "description": "`pending` | `done`",
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticketKey",
          "text",
          "dueAt",
          "state",
          "createdAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Workflow",
      "url": "https://github.com/"
    },
...
    "license": {
      "name": "Proprietary"
    },
    "title": "Workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/hub/board": {
      "get": {
        "operationId": "hubBoard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubBoard"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/hub/board — the composed kanban board + orphan tray.",
        "tags": [
          "hub"
        ]
      }
    },
    "/api/me/hub/inbox": {
      "get": {
        "operationId": "hubInbox",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubInbox"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/hub/inbox — the ranked Needs-you feed (+ morning brief when on).",
        "tags": [
          "hub"
        ]
      }
    },
    "/api/me/hub/links": {
      "delete": {
        "operationId": "hubUnlink",
        "parameters": [
          {
            "description": "Unused; accepted for symmetry",
            "in": "query",
            "name": "ticketKey",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "owner/name",
            "in": "query",
            "name": "repo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "PR target",
            "in": "query",
            "name": "prNumber",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Branch target",
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "DELETE /me/hub/links — remove a manual link.",
        "tags": [
          "hub"
        ]
      },
      "post": {
        "operationId": "hubLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HubLinkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/hub/links — manually link a PR or branch to a ticket.",
        "tags": [
          "hub"
        ]
      }
    },
    "/api/me/hub/runs": {
      "get": {
        "operationId": "hubRuns",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubRuns"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/hub/runs — actions-strip run pills (favorites first-class).",
        "tags": [
          "hub"
        ]
      }
    },
    "/api/me/hub/sync-status": {
      "get": {
        "operationId": "hubSyncStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubSyncStatus"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "hub"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin (the web app proxies /api to this server)",
      "url": "/"
    }
  ],
...
}